Fix CodeQL tainted-numeric-cast alerts: change monopolyAmount from double to int
Agent-Logs-Url: https://github.com/dotoRi-muuk/TRPG-java/sessions/8f4f83d8-0eab-41b1-ac08-df87320d385d

+++ src/main/java/trpg/web/LightningPriestController.java
@@ -40,7 +40,7 @@ public class LightningPriestController {
         public int damageBonus;
         public int finalDamageBonus;
         public boolean monopoly;
-        public double monopolyAmount;
+        public int monopolyAmount;
         public boolean piety;
         public int precision;
         public int chantTurns;
